% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acnh_crafting}
\alias{acnh_crafting}
\title{Animal Crossing: New Horizons crafting materials}
\format{A data frame
\describe{
}}
\source{
\url{https://github.com/chendaniely/animal_crossing}
}
\usage{
acnh_crafting
}
\description{
A dataset containing
the crafting materials, image url, how to obtain, and price
}
\keyword{datasets}
