% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acnh_diy_equipment}
\alias{acnh_diy_equipment}
\title{Animal Crossing: New Horizons equiptment diy recipes}
\format{A data frame
\describe{
}}
\source{
\url{https://github.com/chendaniely/animal_crossing}
}
\usage{
acnh_diy_equipment
}
\description{
A dataset containing
the equiptment, image url, materials needed, size, how it is obtained, and price
}
\keyword{datasets}
