% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acnh_diy_housewares}
\alias{acnh_diy_housewares}
\title{Animal Crossing: New Horizons housewares diy recipes}
\format{A data frame
\describe{
}}
\source{
\url{https://github.com/chendaniely/animal_crossing}

\url{https://github.com/chendaniely/animal_crossing}
}
\usage{
acnh_diy_housewares

acnh_diy_housewares
}
\description{
A dataset containing
the houseware, image url, materials needed, size, how it is obtained, and price

A dataset containing
the houseware, image url, materials needed, size, how it is obtained, and price
}
\keyword{datasets}
