% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acnh_diy_other}
\alias{acnh_diy_other}
\title{Animal Crossing: New Horizons other diy recipes}
\format{A data frame
\describe{
}}
\source{
\url{https://github.com/chendaniely/animal_crossing}
}
\usage{
acnh_diy_other
}
\description{
A dataset containing
the item, image url, materials needed, size, how it is obtained, and price
}
\keyword{datasets}
