% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acnh_diy_wallfloorrug}
\alias{acnh_diy_wallfloorrug}
\title{Animal Crossing: New Horizons wall, floring, and rug diy recipes}
\format{
A data frame
\describe{
}
}
\source{
\url{https://github.com/chendaniely/animal_crossing}
}
\usage{
acnh_diy_wallfloorrug
}
\description{
A dataset containing
the item, image url, materials needed, size, how it is obtained, and price
}
\keyword{datasets}
