% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acnh_fish_n}
\alias{acnh_fish_n}
\title{Animal Crossing: New Horizons fish in the northern hemisphere}
\format{A data frame
\describe{
}}
\source{
\url{https://github.com/chendaniely/animal_crossing}
}
\usage{
acnh_fish_n
}
\description{
A dataset containing
the fish, image url, price, location, time, and month
}
\keyword{datasets}
