% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acnh_fish_s}
\alias{acnh_fish_s}
\title{Animal Crossing: New Horizons fish in the southern hemisphere}
\format{
A data frame
\describe{
}
}
\source{
\url{https://github.com/chendaniely/animal_crossing}
}
\usage{
acnh_fish_s
}
\description{
A dataset containing
the fish, image url, price, location, time, and month
}
\keyword{datasets}
