% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acnh_villagers}
\alias{acnh_villagers}
\title{Animal Crossing: New Horizons villagers}
\format{
A data frame
\describe{
}
}
\source{
\url{https://github.com/chendaniely/animal_crossing}
}
\usage{
acnh_villagers
}
\description{
A dataset containing
the villager, image url, personality, species, birtday, and catchphrase
}
\keyword{datasets}
